clean component
