+++ apps/forms/scaffolds/template-editor/index.ts
@@ -1,0 +1,1 @@
+export * from "./template-editor";
